+++ InventarioSaaS.Domain/Entidades/Producto.cs
@@ -13,7 +13,7 @@ namespace InventarioSaaS.Domain.Entidades
         public required string Nombre { get; set; }
 
         [Required]
-        public required float PrecioVenta { get; set; }
+        public required decimal PrecioVenta { get; set; }
 
         [Required]
         public required int Stock { get; set; }
